backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pdfplumber==0.10.3
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
faiss-cpu==1.13.2
nltk==3.8.1
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.26.2
